+++ apps/web-svelte/src/lib/components/VideoPlayer.svelte
@@ -1068,8 +1068,6 @@
       clearControlsTimer();
       showControls = false;
       closeMenus();
-    } else {
-      surfaceControls();
     }
   }